% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Stations distances from damage.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
   \item{DISTANCE}{distance from fire station in miles}
   \item{DAMAGE}{x 1000 dollars in damage}
}
}
\source{
\url{https://crcpress.com/Statistics-for
-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book
/9781498728850}
}
\usage{
fire
}
\description{
A dataset containing the distance from 15 fire stations
to the location of damage.
}
\keyword{datasets}
