% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Sampling Binomial Experiment}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability of success}
}
\value{
barplot of successes in each sample
}
\description{
Sampling Binomial Experiment
}
\examples{
mybin(iter=100,n=10,p=0.7)
}
