% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Boostrap Confidence Intervals for Sampling}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{The number of iterations taken}

\item{x}{The vector of statistics}

\item{fun}{The function used to manipulate statistics}

\item{alpha}{The percentage from 0 or 100 to determine the interval}

\item{cx}{The x-axis coordinate of a center point}

\item{...}{Any additional number of named or unnamed arguments}
}
\value{
A graphical representation of a bootstrap confidence interval
}
\description{
Boostrap Confidence Intervals for Sampling
}
\examples{
\dontrun{myboot2(x=ddt$DDT)}
}
