% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{Plots of DDT Species}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{The DDT data set.}

\item{SPECIES}{The given name of fish species.}
}
\value{
A plot of of LENGTH v WEIGHT and a named list of the ddt data frame.
}
\description{
Plots of DDT Species
}
\examples{
\dontrun{myddt(df = ddt, SPECIES = "CCATFISH")}
}
