% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Calculation of the Maximum Log Likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The log function to be evaluated using the function}

\item{x}{The vector of values, or a sample}

\item{param}{The paramter of interest; e.g., mean, standard deviation}

\item{...}{Any additional number of named or unnamed arguments}
}
\value{
A summary of slopes, indieces, the parameter of interest, and
a graphical output of where the x and param intersect
}
\description{
Calculation of the Maximum Log Likelihood
}
\examples{
\dontrun{y = c(3,3,4,5)
logbin=function(x,param) log(dbinom(x,prob=param,size=20))
mymaxlik(x=y,param=seq(0,1,length=1000),lfun=logbin}
}
